(DTS Resolver) Proxy system to handle pagination for references, children and parents
